requirements.txt
open-clip-torch
streamlit
chromadb==0.5.0
pillow
numpy==1.26.1
tqdm
opencv-python